+++ Example/Production/govuk_onboarding_ios/ViewControllers/ViewController.swift
@@ -1,19 +1,22 @@
 import UIKit
 import Onboarding
+
 class ViewController: UIViewController {
     override func viewDidLoad() {
         super.viewDidLoad()
         setUpViewController()
     }
+
     func setUpViewController() {
         let onboardingModule =
-        OnboardingModule(dismissAction: {})
+        Onboarding(dismissAction: {})
         let onboardingViewController = onboardingModule.initialViewController(for: .localJSON( "OnboardingResponse"))
         guard let onboardingView = onboardingViewController.view else { return }
         onboardingViewController.didMove(toParent: self)
         view.addSubview(onboardingViewController.view)
         setUpContraints(view: onboardingView)
     }
+
     private func setUpContraints(view: UIView) {
         view.translatesAutoresizingMaskIntoConstraints = false
         NSLayoutConstraint.activate([
